+++ composite/dns-record-check/src/adapter.ts
@@ -9,7 +9,7 @@ const inputParams = {
   record: true,
 }
 
-const execute: ExecuteWithConfig = async (input, config) => {
+const execute: ExecuteWithConfig<Config> = async (input, config) => {
   const validator = new Validator(input, inputParams)
   if (validator.error) throw validator.error