+++ oilpriceapi/src/adapter.ts
@@ -15,7 +15,7 @@ const customError = (data: Record<string, unknown>) => {
   return data.data === null
 }
 
-export const execute: ExecuteWithConfig = async (input, config) => {
+export const execute: ExecuteWithConfig<Config> = async (input, config) => {
   const validator = new Validator(input, customParams)
   if (validator.error) throw validator.error