+++ typings/@chainlink/index.d.ts
@@ -75,9 +75,9 @@ declare module '@chainlink/types' {
 
   export type Execute = (input: AdapterRequest) => Promise<AdapterResponse>
 
-  export type ExecuteWithConfig = (
+  export type ExecuteWithConfig<C extends Config> = (
     input: AdapterRequest,
-    config: Config,
+    config: C,
   ) => Promise<AdapterResponse>
 
   export type ExecuteFactory<C extends Config> = (config?: C) => Execute