+++ wbtc-address-set/src/adapter.ts
@@ -34,7 +34,7 @@ type ChainType = 'btc' | 'eth'
 const inputParams = {}
 
 // Export function to integrate with Chainlink node
-export const execute: ExecuteWithConfig = async (request, config) => {
+export const execute: ExecuteWithConfig<Config> = async (request, config) => {
   const validator = new Validator(request, inputParams)
   if (validator.error) throw validator.error