+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - `dns-query` to query DNS over HTTPS
   - `dns-record-check` to check whether some record provided exists on DNS
   - `outlier-detection`: composite adapter to check for outlier values between multiple sets of data providers
+  - `dydx-stark` to sign the input price data with your private STARK key, and send it to the destination endpoint.
 - Added support for metadata in requests. This gives adapters access to the FM on-chain round state.
 - Moves re-usable test behaviors & testing utils to a new package - `@chainlink/adapter-test-helpers`
 - Added support for using query string parameters as input to adapters.